requirements.txt
streamlit
pandas
numpy
plotly
xarray
cfgrib
ecmwf-opendata